Rename Main page to PackageViewer1 & Add PackageViewer2 page
