+++ src/pages/PackageViewer2/Main.tsx
@@ -1,0 +1,7 @@
+export default function Main() {
+  return (
+    <div>
+      <div className="text-3xl">BlockBlock</div>
+    </div>
+  );
+}
